Move the logic of preview method into exportToPDF
